Resolve issues with multiple properties defined and extending classes
There was an issue with multiple properties being defined on a single
class, this has been resolved.  Additionally if you extended a class in
the local scope it would not resolve correctly, this has been resolved
by passing the object instead of a string representing the path to the
object.  Updated test cases to test for the scenarios.